+++ src/loadweights.cpp
@@ -44,7 +44,7 @@ std::vector<cv::Mat> loadWeightsMat(std::string filename)
 	Weights weights = loadWeights(filename);
 	std::vector<cv::Mat> results;
 
-	for (int i = 0; i < weights.size(); i++) {
+	for (size_t i = 0; i < weights.size(); i++) {
 		const Matrix &matrix = weights[i];
 
 		const int nrows = matrix.size();
